Отрисуй меня полностью 2(реализовал попап)
