+++ js/main.js
@@ -1,2 +1,2 @@
-import './pictures.js';
+import './bigPictures.js';
